requirements.txt
numpy
pandas
matplotlib
python-dotenv
colorama
ta
twelvedata